Update
I changed the web scraper to add in the Discovery and You Are Here Series. This change has caused some other things to change like the naming scheme in the cups JSON file and messing with the data handling. Other than that, I have found some issues for anyone else to work on because school is starting; some cups aren't being scraped? and making the code more efficient.