Wiped Database, Added my Cup Collection, Bug Fixes
Wiped the current database of cups due to it being based on ONLY Been There series cups and the other cup types are not being listed on the GitHub page. I also added my collection of cups. Finally, I changed some debugging messages and messed with the placement of stacked markers.